wip: device-streaming-session
