+++ projects/console-web-server/src/ws/device-streaming/dto/device-streaming.dto.ts
